Batch allocation implemented
Helper functions exist in cluster.py to provision the system.
Tests updated in test_scheduling.py to demonstrate batch provisioning.
batch_planning.py and batch_allocation.py implement the
Planning and Algorithm abstract class respectively.
