+++ topsim/user/plan/batch_planning.py
@@ -18,7 +18,7 @@ import copy
 import networkx as nx
 
 from topsim.core.task import Task
-from topsim.core.planning import Planning
+from topsim.algorithms.planning import Planning
 from topsim.core.planner import WorkflowStatus, WorkflowPlan
