+++ topsim/user/schedule/dynamic_plan.py
@@ -16,7 +16,7 @@ import copy
 import logging
 import pandas as pd
 
-from topsim.core.algorithm import Algorithm
+from topsim.algorithms.scheduling import Algorithm
 from topsim.core.planner import WorkflowStatus
 from topsim.core.task import TaskStatus
